refactor CalendarUnit and AlignmentReference class
