+++ backend/src/main/java/com/bakdata/conquery/models/forms/util/CalendarUnit.java
@@ -1,0 +1,14 @@
+package com.bakdata.conquery.models.forms.util;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum CalendarUnit {
+	DAYS(Alignment.DAY),
+	QUARTERS(Alignment.QUARTER),
+	YEARS(Alignment.YEAR);
+
+	@Getter
+	private final Alignment alignment;
+}
